change to dialog
